Bring back request module for file uploads
I could not get file uploads to work with axios, even though some examples in the internet seemed to claim that axios does support file uploads in node. In my testing, I could not get it to work, however, which is why I decided to pull back in the request module.